% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kdeborder.R
\docType{package}
\name{kdeborder}
\alias{kdeborder}
\title{Kernel density estimation correcting for border bias}
\description{
This package proposes some R codes to compute the kernel density estimates of
two-dimensional data points, using an extension of Ripley's circumference
method to correct for border bias.
}
\author{
Arthur Charpentier and Ewen Gallic \email{ewen.gallic@gmail.com}
}
