% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kdeborder.R
\name{sCircle}
\alias{sCircle}
\title{Creates "n" points on a circle centered in "centre" with a radius of "radius"}
\usage{
sCircle(n = 100, centre = c(0, 0), radius)
}
\arguments{
\item{n}{number of points to define the circle}

\item{centre}{center of the circle}

\item{radius}{radius of the circle}
}
\value{
Returns a matrix with two columns corresponding to the (x,y) coordinates of the circle of centre `centre`
  and radius `radius`. The number of lines depends on the number of points asked using the parameter `n`.
}
\description{
Creates "n" points on a circle centered in "centre" with a radius of "radius"
}
\examples{
sCircle(n=100, centre = c(0,0), radius = 1)
}
