% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sKDE}
\alias{sKDE}
\title{Computes an estimation of the density using Kernel Density estimator,
correcting for fontier effects.}
\usage{
sKDE(U, polygon, optimal = TRUE, h = 0.1, parallel = FALSE, n_clusters = 4)
}
\arguments{
\item{U}{data points.}

\item{polygon}{polygon (simple figure geometry created with {sf}) on which points lie.}

\item{optimal}{if TRUE, uses Hpi() to select the optimal bandwidth.}

\item{h}{only if optimal=FALSE, scalar bandwidth.}

\item{parallel}{if TRUE, computes the weights using clusters.}

\item{n_clusters}{only if n_clusters=TRUE, defines the number of clusters. (Set to NULL for automatic detection (on Unix)).}
}
\value{
Returns a list whose elements are:
  > X:    x coordinates at wich estimate is evaluated,
  > Y:    y coordinates at wich estimate is evaluated,
  > Z:    density estimates,
  > ZNA:  density estimates with NA values for points outside the polygon,
  > H:    bandwidth matrix,
  > W:    vector of weights.
}
\description{
Computes an estimation of the density using Kernel Density estimator,
correcting for fontier effects.
}
\examples{
data(acci)
# Estimation with correction
polygon_finistere <-
cbind(
  c(acci$finistere$polygon$long, acci$finistere$polygon$long[1]),
  c(acci$finistere$polygon$lat, acci$finistere$polygon$lat[1])
) \%>\%
  list() \%>\%
  sf::st_polygon()
smoothed_fin <- sKDE(U = acci$finistere$points, polygon = polygon_finistere,
optimal=TRUE, parallel = FALSE)
}
\references{
Charpentier, A. & Gallic, E. (2015). Kernel density estimation based on Ripley’s correction. GeoInformatica, 1-22.
(\href{https://link.springer.com/article/10.1007/s10707-015-0232-z}{Springer})
}
\seealso{
\code{\link{sWeights}} which this function wraps
}
