% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sKDE_without_c}
\alias{sKDE_without_c}
\title{Computes an estimation of the density using Kernel Density estimator,
without correcting for fontier effects.}
\usage{
sKDE_without_c(U, polygon, optimal = TRUE, h = 0.1)
}
\arguments{
\item{U}{data points.}

\item{polygon}{polygon (simple figure geometry created with {sf}) on which points lie.}

\item{optimal}{if TRUE, uses Hpi() to select the optimal bandwidth.}

\item{h}{only if optimal=FALSE, scalar bandwidth.}
}
\value{
Returns a list whose elements are:
  > X:    x coordinates at wich estimate is evaluated,
  > Y:    y coordinates at wich estimate is evaluated,
  > Z:    density estimates,
  > ZNA:  density estimates with NA values for points outside the polygon,
  > H:    bandwidth matrix,
  > W:    vector of weights.
}
\description{
Computes an estimation of the density using Kernel Density estimator,
without correcting for fontier effects.
}
\examples{
data(acci)
polygon_finistere <-
cbind(
  c(acci$finistere$polygon$long, acci$finistere$polygon$long[1]),
  c(acci$finistere$polygon$lat, acci$finistere$polygon$lat[1])
) \%>\%
  list() \%>\%
  sf::st_polygon()
smoothed_fin_nc <- sKDE_without_c(U = acci$finistere$points,
    polygon = polygon_finistere, optimal=TRUE)
}
